Restore engine/lsm/errors.go and diagnostics.go as standalone files
Sentinels and the Diagnostics surface read better as their own files
(conventional Go layout, easier to navigate) even though they have no
direct _test.go pair: errors.go is exercised through every error
return; diagnostics.go is exercised through Diagnostics-consuming
tests in lsm_test.go.

errors.go: ErrFillTables, ErrMemtableNotInitialized, ErrFlushNilMemtable,
  ErrLSMNil*, ErrLSMClosed.
diagnostics.go: CompactionDiagnostics / RangeFilterDiagnostics /
  Diagnostics types, LSM.Diagnostics(), countMemIndexEntries,
  rangeFilterMetrics + record/snapshot helpers, levelManager.entryCount,
  levelHandler.tablesSnapshot.
